RavenDB-24211 Add chat trimming section and UI improvements
