+++ src/Raven.Studio/typescript/components/pages/database/aiHub/aiAgents/edit/utils/editAiAgentUtils.ts
@@ -1,5 +1,6 @@
 import genUtils from "common/generalUtils";
-import { EditAiAgentFormData } from "./editAiAgentValidation";
+import { AiAgentTrimmingMethod, EditAiAgentFormData } from "./editAiAgentValidation";
+import { TimeInSeconds } from "common/constants/timeInSeconds";
 
 function mapFromDto(
     dto?: Raven.Client.Documents.Operations.AI.Agents.AiAgentConfiguration,
@@ -16,12 +17,25 @@ function mapFromDto(
             outputSchema: "",
             isEnableDocumentExpiration: !isDocumentExpirationEnabled,
             persistenceConversationIdPrefix: "",
-            persistenceExpiresInSeconds: 2592000, // 30 days
+            persistenceExpiresInSeconds: TimeInSeconds.Day * 30,
+            parameterInput: "",
             parameters: [],
             queries: [],
             actions: [],
+            maxModelIterationsPerCall: null,
             testPrompt: "",
             testParameters: [],
+            trimming: {
+                method: null,
+                historyExpirationInSeconds: null,
+                messagesLengthBeforeTruncate: null,
+                messagesLengthAfterTruncate: null,
+                maxTokensBeforeSummarization: null,
+                maxTokensAfterSummarization: null,
+                resultPrefix: "",
+                summarizationTaskBeginningPrompt: "",
+                summarizationTaskEndPrompt: "",
+            },
         };
     }
 
@@ -35,6 +49,7 @@ function mapFromDto(
         isEnableDocumentExpiration: !isDocumentExpirationEnabled,
         persistenceConversationIdPrefix: dto.Persistence.ConversationIdPrefix,
         persistenceExpiresInSeconds: dto.Persistence.Expires ? Number(dto.Persistence.Expires) / 1000 : null, // Expires is in milliseconds
+        parameterInput: "",
         parameters:
             dto.Parameters?.map((x) => ({
                 name: x,
@@ -58,6 +73,20 @@ function mapFromDto(
                 isSaved: true,
                 isEditing: false,
             })) ?? [],
+        maxModelIterationsPerCall: dto.MaxModelIterationsPerCall,
+        trimming: {
+            method: getTrimmingMethod(dto),
+            historyExpirationInSeconds: dto.ChatReduction?.History?.HistoryExpiration
+                ? Number(dto.ChatReduction.History.HistoryExpiration) / 1000
+                : null,
+            messagesLengthBeforeTruncate: dto.ChatReduction?.Truncate?.MessagesLengthBeforeTruncate,
+            messagesLengthAfterTruncate: dto.ChatReduction?.Truncate?.MessagesLengthAfterTruncate,
+            maxTokensBeforeSummarization: dto.ChatReduction?.Tokens?.MaxTokensBeforeSummarization,
+            maxTokensAfterSummarization: dto.ChatReduction?.Tokens?.MaxTokensAfterSummarization,
+            resultPrefix: dto.ChatReduction?.Tokens?.ResultPrefix,
+            summarizationTaskBeginningPrompt: dto.ChatReduction?.Tokens?.SummarizationTaskBeginningPrompt,
+            summarizationTaskEndPrompt: dto.ChatReduction?.Tokens?.SummarizationTaskEndPrompt,
+        },
         testPrompt: "",
         testParameters:
             dto.Parameters?.map((x) => ({
@@ -65,6 +94,18 @@ function mapFromDto(
                 value: "",
             })) ?? [],
     };
+}
+
+function getTrimmingMethod(
+    dto?: Raven.Client.Documents.Operations.AI.Agents.AiAgentConfiguration
+): AiAgentTrimmingMethod {
+    if (dto.ChatReduction?.Tokens) {
+        return "Tokens";
+    }
+    if (dto.ChatReduction?.Truncate) {
+        return "Truncate";
+    }
+    return null;
 }
 
 function mapToDto(
@@ -85,6 +126,7 @@ function mapToDto(
                     ? genUtils.formatAsTimeSpan(formData.persistenceExpiresInSeconds * 1000)
                     : null,
         },
+        Parameters: formData.parameters?.map((x) => x.name) ?? [],
         Queries:
             formData.queries?.map((x) => ({
                 Name: x.name,
@@ -100,9 +142,31 @@ function mapToDto(
                 ParametersSampleObject: x.parametersSampleObject || null,
                 ParametersSchema: x.parametersSchema || null,
             })) ?? [],
-        Parameters: formData.parameters?.map((x) => x.name) ?? [],
-        ChatTrimming: undefined, // omit for now - lets wait for the design
-        MaxModelIterationsPerCall: undefined, // omit for now - lets wait for the design
+        MaxModelIterationsPerCall: formData.maxModelIterationsPerCall,
+        ChatReduction: {
+            History: formData.trimming?.historyExpirationInSeconds
+                ? {
+                      HistoryExpiration: genUtils.formatAsTimeSpan(formData.trimming.historyExpirationInSeconds * 1000),
+                  }
+                : null,
+            Tokens:
+                formData.trimming?.method === "Tokens"
+                    ? {
+                          MaxTokensBeforeSummarization: formData.trimming.maxTokensBeforeSummarization,
+                          MaxTokensAfterSummarization: formData.trimming.maxTokensAfterSummarization,
+                          ResultPrefix: formData.trimming.resultPrefix,
+                          SummarizationTaskBeginningPrompt: formData.trimming.summarizationTaskBeginningPrompt,
+                          SummarizationTaskEndPrompt: formData.trimming.summarizationTaskEndPrompt,
+                      }
+                    : null,
+            Truncate:
+                formData.trimming?.method === "Truncate"
+                    ? {
+                          MessagesLengthBeforeTruncate: formData.trimming.messagesLengthBeforeTruncate,
+                          MessagesLengthAfterTruncate: formData.trimming.messagesLengthAfterTruncate,
+                      }
+                    : null,
+        },
     };
 }
