+++ src/Raven.Studio/typescript/components/pages/database/aiHub/aiAgents/utils/aiAgentsUtils.ts
@@ -1,14 +1,14 @@
 import moment from "moment";
 import { AiAgentDocMessage, AiAgentMessage } from "./aiAgentsTypes";
 
-const getContentFromDoc = (docMessage: AiAgentDocMessage): string => {
+function getContentFromDoc(docMessage: AiAgentDocMessage): string {
     if (docMessage.content && (docMessage.role === "assistant" || docMessage.role === "tool")) {
         return JSON.stringify(JSON.parse(docMessage.content), null, 2);
     }
     return docMessage.content;
-};
+}
 
-const mapMessageFromDoc = (docMessage: AiAgentDocMessage): AiAgentMessage => {
+function mapMessageFromDoc(docMessage: AiAgentDocMessage): AiAgentMessage {
     return {
         id: _.uniqueId(),
         role: docMessage.role,
@@ -24,7 +24,7 @@ const mapMessageFromDoc = (docMessage: AiAgentDocMessage): AiAgentMessage => {
         date: docMessage.date ? moment(docMessage.date).format(aiAgentsUtils.messageDateFormat) : null,
         usage: docMessage.usage,
     };
-};
+}
 
 export const aiAgentsUtils = {
     mapMessageFromDoc,
